restructured code into modules, improved summarization code by tuning length of summary to be proportional to lenth of full text
